feat: move user creation query to firestore
