+++ src/app/(dashboard)/create/page.tsx
@@ -3,14 +3,14 @@
 import { useRouter } from 'next/navigation';
 import { Popup } from 'components/Popup';
 import { UserForm } from 'components/UserForm';
-import { addNewUser } from 'lib/firebase/helpers';
 import { User } from '@/types/user';
+import { addNewUser } from 'lib/data';
 
 export default function CreateUserPage() {
   const router = useRouter();
 
-  const handleSubmit = (userData: User) => {
-    addNewUser(userData);
+  const handleSubmit = async (userData: User) => {
+    await addNewUser(userData);
 
     router.push('/');
   };
