+++ src/lib/data/index.ts
@@ -1,6 +1,13 @@
-import { collection, getDocs, orderBy, query } from 'firebase/firestore';
+import {
+  addDoc,
+  collection,
+  getDocs,
+  orderBy,
+  query,
+  serverTimestamp,
+} from 'firebase/firestore';
 import { newDb } from 'lib/firebase';
-import { Users } from '@/types/user';
+import { User, Users } from '@/types/user';
 import { convertUnixToDateFormat } from 'lib/helpers/convertUnixToDateFormat';
 
 export enum OrderByDirection {
@@ -8,8 +15,10 @@ export enum OrderByDirection {
   Desc = 'desc',
 }
 
+const userCollection = collection(newDb, 'users');
+
 export async function fetchUsers(field: string, direction: OrderByDirection) {
-  const q = query(collection(newDb, 'users'), orderBy(field, direction));
+  const q = query(userCollection, orderBy(field, direction));
   const querySnapshot = await getDocs(q);
 
   const docs = querySnapshot.docs;
@@ -19,4 +28,12 @@ export async function fetchUsers(field: string, direction: OrderByDirection) {
   })) as Users;
 
   return users;
+}
+
+export async function addNewUser(user: User) {
+  const docRef = await addDoc(userCollection, {
+    ...user,
+    createdAt: serverTimestamp(),
+  });
+  // TODO: Implement success notification with docRef.id (id of successfully created user
 }
